aead: Move pdding to streaming
