+++ src/hazardous/aead/chacha20poly1305.rs
@@ -124,23 +124,6 @@ pub(crate) fn poly1305_key_gen(
     OneTimeKey::from_slice(&tmp_buffer[..POLY1305_KEYSIZE]).unwrap()
 }
 
-#[inline]
-/// Padding size that gives the needed bytes to pad `input` to an integral
-/// multiple of 16.
-pub(crate) fn padding(input: usize) -> usize {
-    if input == 0 {
-        return 0;
-    }
-
-    let rem = input % 16;
-
-    if rem != 0 {
-        16 - rem
-    } else {
-        0
-    }
-}
-
 /// Authenticates the ciphertext, ad and their lengths.
 fn process_authentication(
     auth_ctx: &mut Poly1305,
@@ -269,54 +252,6 @@ mod public {
 #[cfg(test)]
 mod private {
     use super::*;
-
-    mod test_padding {
-        use super::*;
-        #[test]
-        fn test_length_padding() {
-            assert_eq!(padding(0), 0);
-            assert_eq!(padding(1), 15);
-            assert_eq!(padding(2), 14);
-            assert_eq!(padding(3), 13);
-            assert_eq!(padding(4), 12);
-            assert_eq!(padding(5), 11);
-            assert_eq!(padding(6), 10);
-            assert_eq!(padding(7), 9);
-            assert_eq!(padding(8), 8);
-            assert_eq!(padding(9), 7);
-            assert_eq!(padding(10), 6);
-            assert_eq!(padding(11), 5);
-            assert_eq!(padding(12), 4);
-            assert_eq!(padding(13), 3);
-            assert_eq!(padding(14), 2);
-            assert_eq!(padding(15), 1);
-            assert_eq!(padding(16), 0);
-        }
-
-        // Proptests. Only executed when NOT testing no_std.
-        #[cfg(feature = "safe_api")]
-        mod proptest {
-            use super::*;
-
-            quickcheck! {
-                // The usize that padding() returns should always
-                // be what remains to make input a multiple of 16 in length.
-                fn prop_padding_result(input: usize) -> bool {
-                    let rem = padding(input);
-
-                    ((input + rem) % 16) == 0
-                }
-            }
-
-            quickcheck! {
-                // padding() should never return a usize above 15.
-                // The usize must always be in range of 0..=15.
-                fn prop_result_never_above_15(input: usize) -> bool {
-                    padding(input) < 16
-                }
-            }
-        }
-    }
 
     mod test_process_authentication {
         use super::*;
